pyproject.toml
[project]
name = "mla-trust"
version = "0.1.0"
description = "mla-trust"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate==1.5.2",
    "aioconsole==0.8.1",
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.12",
    "aiosignal==1.3.2",
    "airportsdata==20250224",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "astor==0.8.1",
    "async-timeout==5.0.1",
    "attrdict>=2.0.1",
    "attrs==25.1.0",
    "audioread==3.0.1",
    "backoff==2.2.1",
    "beautifulsoup4==4.13.3",
    "black==25.1.0",
    "blake3==1.0.4",
    "bs4==0.0.2",
    "build==1.2.2.post1",
    "cachetools==5.5.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cloudpickle==3.1.1",
    "compressed-tensors==0.9.1",
    "cupy-cuda12x==13.3.0",
    "datasets==3.2.0",
    "decorator==5.2.1",
    "depyf==0.18.0",
    "dill==0.3.8",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "einops==0.8.1",
    "einx==0.3.0",
    "email-validator==2.2.0",
    "encodec==0.1.1",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.9",
    "fastapi-cli==0.0.7",
    "fastrlock==0.8.3",
    "filelock==3.17.0",
    "frozendict==2.4.6",
    "frozenlist==1.5.0",
    "fsspec==2024.9.0",
    "ftfy==6.3.1",
    "gguf==0.10.0",
    "gmpy2==2.2.1",
    "google-ai-generativelanguage==0.6.2",
    "google-api-core==2.24.1",
    "google-api-python-client==2.161.0",
    "google-auth==2.38.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.5.2",
    "googleapis-common-protos==1.67.0",
    "greenlet==3.1.1",
    "grpcio==1.70.0",
    "grpcio-status==1.63.0rc1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "huggingface-hub==0.28.1",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "iniconfig==2.0.0",
    "intel-cmplr-lib-ur==2025.0.4",
    "intel-openmp==2025.0.4",
    "interegular==0.3.3",
    "jinja2==3.1.6",
    "jiter==0.8.2",
    "joblib==1.4.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "lark==1.2.2",
    "lazy-loader==0.4",
    "librosa==0.11.0",
    "litellm==1.35.32",
    "llvmlite==0.43.0",
    "lm-format-enforcer==0.10.11",
    "lxml==5.3.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mistral-common==1.5.3",
    "mkl==2025.0.1",
    "mkl-fft==1.3.11",
    "mkl-random==1.2.8",
    "mkl-service==2.4.0",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "msgspec==0.19.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numba==0.60.0",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "open-clip-torch==2.31.0",
    "openai==1.65.1",
    "opencv-python-headless==4.11.0.86",
    "outlines==0.1.11",
    "outlines-core==0.1.26",
    "packaging==24.2",
    "pandas==2.2.3",
    "partial-json-parser==0.2.1.1.post5",
    "pathspec==0.12.1",
    "peft>=0.15.2",
    "pillow==11.1.0",
    "pip==25.0.1",
    "pip-tools==7.4.1",
    "platformdirs==4.3.6",
    "playwright==1.50.0",
    "pluggy==1.5.0",
    "pooch==1.8.2",
    "prometheus-client==0.21.1",
    "prometheus-fastapi-instrumentator==7.0.2",
    "propcache==0.2.1",
    "proto-plus==1.26.0",
    "protobuf==4.25.6",
    "psutil==7.0.0",
    "py-cpuinfo==9.0.0",
    "pyarrow==19.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pybind11==2.13.6",
    "pycountry==24.6.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pyee==12.1.1",
    "pygments==2.19.1",
    "pyparsing==3.2.1",
    "pyproject-hooks==1.2.0",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "pyzmq==26.2.1",
    "qwen-vl-utils>=0.0.10",
    "ray==2.40.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==13.9.4",
    "rich-toolkit==0.13.2",
    "rpds-py==0.23.1",
    "rsa==4.9",
    "ruff==0.9.7",
    "safetensors==0.5.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==3.4.1",
    "sentencepiece==0.2.0",
    "setuptools==77.0.3",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "soupsieve==2.6",
    "soxr==0.5.0.post1",
    "starlette==0.45.3",
    "sympy==1.13.1",
    "tbb==2022.0.0",
    "tcmlib==1.2.0",
    "threadpoolctl==3.5.0",
    "tiktoken==0.9.0",
    "timm==1.0.15",
    "tokenizers==0.21.0",
    "toml==0.10.2",
    "tomli==2.2.1",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "triton==3.1.0",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "umf==0.9.1",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "vector-quantize-pytorch==1.22.3",
    "vllm==0.7.3",
    "vocos==0.1.0",
    "watchfiles==1.0.4",
    "wcwidth==0.2.13",
    "websockets==15.0",
    "wheel==0.45.1",
    "xformers==0.0.28.post3",
    "xgrammar==0.1.11",
    "xxhash==3.5.0",
    "yarl==1.18.3",
    "zipp==3.21.0",
]

[project.optional-dependencies]
flash-attn = [
  "flash-attn==2.7.4.post1"
]

[tool.uv]
no-build-isolation-package = ["flash-attn"]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[tool.ruff]
# In addition to the standard set of exclusions, omit all tests, plus a specific file.
extend-exclude = ["models/src"]
